metrics_pillars_of_observability_tutorial/requirements.txt
Flask==2.2.2
prometheus-client==0.16.0
opentelemetry-api==1.15.0
opentelemetry-exporter-prometheus==1.12.0rc1
opentelemetry-sdk==1.15.0
